kboot: Rework how kboot_prepare_fw() works
Now it just builds and returns the firmware portion of the initramfs. It
is up to the caller to prepend it to the user-supplied initramfs, unless
no initramfs was set, in which case kboot_prepare_dt will implicitly
call kboot_prepare_fw.

Then make the payload code and linux.py handle this properly.

+++ src/cpio.h
@@ -12,7 +12,7 @@ struct cpio *cpio_init(void);
 int cpio_add_file(struct cpio *c, const char *name, const u8 *bfr, size_t sz);
 int cpio_add_dir(struct cpio *c, const char *name);
 size_t cpio_get_size(struct cpio *c);
-size_t cpio_finalize(struct cpio *c, u8 *bfr, size_t bfr_size);
+int cpio_finalize(struct cpio *c, u8 *bfr, size_t bfr_size);
 void cpio_free(struct cpio *c);
 
 #endif
